Added ability to cut data
We should now have the ability to cut data on any parameter given in the
shotData. Cuts can be set easily using the cutter GUI which sets the
cutTable located in the base workspace. I've modified liveAnalysis to
call cutDataFunc every cycle which returns the shotData with the cuts
specified in the cutTable.