Added live image viewer to liveAnalysis
I've added a set of axes to liveAnalysis which allows you to preview
images from the shotData. You can choose to track the most recent shot
loaded, pick a shot to preview from an index number or choose to display
the image related to a data point selected from the primary axes.