Fleshed out setlist control
I've added a setlist control window which can be opened from
liveAnalysis. I've also added a mulligan button in liveAnalysis which
allows you to mulligan shots as selected from the liveAnalysis plot. So
far the setlist control window only allows you to set instant variables
and mulligan shots as Matlab doesn't have the innate ability in the GUI
to have 3d matricies so doing groups for sequence variables is a little
tricky so I'll leave this for later implementation.