Autorun update
I've updated the try, catch at the end of autorun that tries to stick
the processed data into base shotData. Originally the try block had a
bunch of stuff in it which can also throw errors and cause problems
(which happened yesterday). Now the try block only tries to import
shotData from the base workspace and everything else is in an
independent if block which only runs if shotData was succesfully
imported