Added version number
From now on I think it would be a good idea to have some kind of version
numbering implemented. The first digit will be incremented when we do
some change to the code which breaks compatibility with other programs
like FleaRbRy. The second digit will be incremented when new features
are added and the final digit will be incremented for small bug fixes.

As this version seems to be working I'll go ahead and start the version
number at 1.0.0