Modified autorun and reloadDataFile
I've broken out autorun into autorun and a shotProcesser function (which
does the heavy lifting of processing shots). This has allowed me to
update reloadDataFile to do parallel processing of shots when the user
requests multiple shots to be loaded. This should improve the speed at
which old shots can be loaded.