Reworked how liveAnalysis updates
liveAnalysis now has a hidden variable in its handles, requiresUpdate,
which when set to true will cause the GUI to update its internal
shotData. Previously the internal shotData was updated every third of a
second which was computationally expensive. I've updated all the
functions which alter the base shotData or cutTable to update
liveAnalysis each time a change is made.