Added time tagger OD measurement class
I've added a class for working out the OD for the count data taken from
the time tagger. I've updated the shotProcessor and liveAnalysis to
accomodate the new class. Additionally i've changed
basicFittingClass->basicImageFittingClass to reflect the fact that it's
properties are only fundamental to classes that do image fitting.