Fixing issue with time tagger processing
Matlab was casting the iterator in a for loop as an 8 bit int which was
then multplied by 2 later on. This caused problems when the iterator
became greater than 63, i.e. if we averaged over more than 21 shots.
I've solved this by deliberately casting the iterator as an unsigned 16
bit integer