Added running G2 calculation class
This class is for doing running G2 calculations, i.e. built up over many
shots. The class forms a persistent object (in that it saves itself to
the workspace and loads from there) which allows one to see the build up
of g2 over multiple shotfiles (as opposed to working out g2 for one file
that contains multiple shots). There doesn't seem to be any need for
liveAnalysis so I've bypassed this and just got the shotProcessor to
replot g2 after each new file is processed